Add type alias Square
And define it as i8.  This provides a single definition of the Square
type, making the type definitions more clear.  Also, since it is
signed, it eliminates the need to do casting when adding directions.
The drawback is that casting is needed when doing array indexing, but
that seems to be more consistent with the conceptual "square" type
anyway.

Another advantage is that squares are now represented using only 1
byte, whereas previously 8 bytes were being used (usize).

+++ src/pieces.rs
@@ -106,10 +106,10 @@ pub const WHITE: usize = 0;
 pub const BLACK: usize = 1;
 pub const BOTH: usize = 2;
 
-pub const KNIGHT_DIR: [i32; 8] = [-8, -19, -21, -12, 8, 19, 21, 12];
-pub const ROOK_DIR: [i32; 4] = [-1, -10, 1, 10];
-pub const BISHOP_DIR: [i32; 4] = [-9, -11, 11, 9];
-pub const KING_DIR: [i32; 8] = [-1, -10, 1, 10, -9, -11, 11, 9];
+pub const KNIGHT_DIR: [i8; 8] = [-8, -19, -21, -12, 8, 19, 21, 12];
+pub const ROOK_DIR: [i8; 4] = [-1, -10, 1, 10];
+pub const BISHOP_DIR: [i8; 4] = [-9, -11, 11, 9];
+pub const KING_DIR: [i8; 8] = [-1, -10, 1, 10, -9, -11, 11, 9];
 
 // SLIDERS[color] produces an array that can be iterated through
 pub const SLIDERS: [[Piece; 3]; 2] = [[Piece::WB, Piece::WR, Piece::WQ], [Piece::BB, Piece::BR, Piece::BQ]];
@@ -121,7 +121,7 @@ pub const NON_SLIDERS: [[Piece; 2]; 2] = [[Piece::WN, Piece::WK], [Piece::BN, Pi
 // piece.  A zero value is used to indicate the end, since the counts
 // are not the same.  Pawns are not included.  Storing with vectors
 // would make more sense but can't be statically allocated.
-pub const DIRECTIONS: [[i32; 9]; 13] =
+pub const DIRECTIONS: [[i8; 9]; 13] =
     [ [0; 9],
        [0; 9],
        [ -8, -19, -21, -12, 8, 19, 21, 12, 0 ],
